test: add missing tests
